cosmetic

+++ source/LGList.pas
@@ -599,9 +599,6 @@ type
     function  NearestGT(constref aValue: T): SizeInt;
     function  LeftmostGE(constref aValue: T): SizeInt;
     function  SelectDistinctArray(constref a: array of T): TArray;
-    function  IndexInRange(aIndex: SizeInt): Boolean; inline;
-    procedure CheckIndexRange(aIndex: SizeInt); inline;
-    procedure IndexOutOfBoundError(aIndex: SizeInt); inline;
     function  GetHeadEnumerator(aHighBound: SizeInt): THeadEnumerator; inline;
     function  GetTailEnumerator(aLowBound: SizeInt): TTailEnumerator; inline;
     function  GetRangeEnumerator(aLowBound, aHighBound: SizeInt): TTailEnumerator; inline;
@@ -3199,14 +3196,18 @@ end;
 
 function TGLiteComparableSortedList.GetItem(aIndex: SizeInt): T;
 begin
-  CheckIndexRange(aIndex);
-  Result := FBuffer.FItems[aIndex];
+  if SizeUInt(aIndex) < SizeUInt(Count) then
+    Result := FBuffer.FItems[aIndex]
+  else
+    raise ELGListError.CreateFmt(SEIndexOutOfBoundsFmt, [aIndex]);
 end;
 
 procedure TGLiteComparableSortedList.SetItem(aIndex: SizeInt; aValue: T);
 begin
-  CheckIndexRange(aIndex);
-  DoSetItem(aIndex, aValue);
+  if SizeUInt(aIndex) < SizeUInt(Count) then
+    DoSetItem(aIndex, aValue)
+  else
+    raise ELGListError.CreateFmt(SEIndexOutOfBoundsFmt, [aIndex]);
 end;
 
 procedure TGLiteComparableSortedList.DoSetItem(aIndex: SizeInt; const aValue: T);
@@ -3376,22 +3377,6 @@ begin
   System.SetLength(Result, Succ(I));
 end;
 
-function TGLiteComparableSortedList.IndexInRange(aIndex: SizeInt): Boolean;
-begin
-  Result := (aIndex >= 0) and (aIndex < Count);
-end;
-
-procedure TGLiteComparableSortedList.CheckIndexRange(aIndex: SizeInt);
-begin
-  if not IndexInRange(aIndex) then
-    IndexOutOfBoundError(aIndex);
-end;
-
-procedure TGLiteComparableSortedList.IndexOutOfBoundError(aIndex: SizeInt);
-begin
-  raise ELGListError.CreateFmt(SEIndexOutOfBoundsFmt, [aIndex]);
-end;
-
 function TGLiteComparableSortedList.GetHeadEnumerator(aHighBound: SizeInt): THeadEnumerator;
 begin
   Result.Init(Self, aHighBound);
@@ -3559,13 +3544,15 @@ end;
 
 procedure TGLiteComparableSortedList.Delete(aIndex: SizeInt);
 begin
-  CheckIndexRange(aIndex);
-  DeleteItem(aIndex);
+  if SizeUInt(aIndex) < SizeUInt(Count) then
+    DeleteItem(aIndex)
+  else
+    raise ELGListError.CreateFmt(SEIndexOutOfBoundsFmt, [aIndex]);
 end;
 
 function TGLiteComparableSortedList.TryDelete(aIndex: SizeInt): Boolean;
 begin
-  Result := IndexInRange(aIndex);
+  Result := SizeUInt(aIndex) < SizeUInt(Count);
   if Result then
     DeleteItem(aIndex);
 end;
